move MessageInteraction into Message
